Maybe this will work

+++ rewrite-gradle-tooling-model/model/src/main/java/org/openrewrite/gradle/toolingapi/OpenRewriteModelBuilder.java
@@ -122,7 +122,9 @@ public class OpenRewriteModelBuilder {
                 return OpenRewriteModel.from(customModelBuilder.get());
             } finally {
                 try {
-                    Files.delete(init);
+                    if (Files.exists(init)) {
+                        Files.delete(init);
+                    }
                 } catch (IOException e) {
                     //noinspection ThrowFromFinallyBlock
                     throw new UncheckedIOException(e);
